Split networking functions to proper components
