--- source/network/ncprot/src/ncprot_message.cpp
@@ -1,8 +1,0 @@
-#include "ncprot_message.h"
-
-std::byte const* mel::ncprot::message::data() const noexcept
-{
-    return data_.data();
-}
-
-std::size_t mel::ncprot::message::size() const noexcept { return data_.size(); }
