+++ source/network/ncprot/src/ncprot_types.cpp
@@ -1,0 +1,1 @@
+#include "ncprot_types.h"
