+++ source/network/ncprot/test/ncprot_serialization.t.cpp
@@ -1,0 +1,3 @@
+#include <catch2/catch.hpp>
+
+#include "ncprot_serialization.h"
